Time: 0 ms (100%), Space: 30.3 MB (85.58%) - LeetHub

+++ 3903-smallest-stable-index-i/3903-smallest-stable-index-i.cpp
@@ -4,8 +4,8 @@ public:
         int i = 0;
         int j = 0;
         int m = nums.size() - 1;
-        long long max_unit;
-        long long min_unit;
+        int max_unit ;
+        int min_unit ;
         int t;
 
         // Loop through each possible index i
@@ -13,13 +13,13 @@ public:
             // Find max from 0 to i
             max_unit = nums[0];
             for (j = 0; j <= i; j++) {
-                max_unit = max(max_unit, (long long)nums[j]);
+                max_unit = max(max_unit, nums[j]);
             }
             
             // Find min from i to m (end of array)
             min_unit = nums[i];
             for (j = i; j <= m; j++) {
-                min_unit = min(min_unit, (long long)nums[j]);
+                min_unit = min(min_unit, nums[j]);
             }
 
             // Calculate instability score
